+++ quantify_scheduler/backends/graph_compilation.py
@@ -189,6 +189,11 @@ class DeviceCompilationConfig(DataStructure):
                 "compilation_func": "quantify_scheduler.backends.circuit_to_device."
                 + "set_pulse_and_acquisition_clock",
             },
+            # TODO in SE-540.
+            # {
+            #    "name": "insert_compensating_pulse",
+            #    "compilation_func": to_be_determined
+            # }
             {
                 "name": "determine_absolute_timing",
                 "compilation_func": "quantify_scheduler.compilation._determine_absolute_timing",